Add GH Action to tag PKGBUILD changes
To update a given Arch Linux package you update the PKGBUILD file.
Mainly you update the `pkgver` variable, but the version is somewhat
more complicated than that. You can also signal build-specific changes
with the variable `pkgrel`, which is exclusive to build iterations of
Arch Linux packages and is not related to the source's version. This
is done so if the source doesn't change, but you need to change some
build step for any given reason, you'd update `pkgrel` while keeping
the `pkgver` (which represents source version) the same.

However, it's even more complicated than that. If you're curious:
https://wiki.archlinux.org/index.php/PKGBUILD#Version

The main complication for me here is that PKGBUILD files can specify
a `pkgver` function that will be run when building, which will
dynamically change the `pkgver` variable value. This is done mostly to
rewrite dashes `-` as underscores `_` since dashes are invalid in
PKGBUILD versions.

But, that also means that I can't know the final version until I try
to build the package and the `pkgver` function, if any, runs and
updates the written version with the generated one.

An easy solution is to simply build a package, and from the resulting
package, make use of either the `.PKGINFO` or `.BUILDINFO` files which
will contain a representation of the final metadata for the package.

This way, I can have GitHub Actions automatically figure out which
version any new push to the repository deploys, and then create and
push a tag back to the repository with that tag, automating the work
I'd do manually before.

One important thing to note here, however, is that events (such as
pushes) from within a GitHub Action **do not** trigger further events
when using the `GITHUB_TOKEN` authentication. This is needed for me
since I want to build and deploy releases on tag pushes.

Currently the only solution is to generated a Personal Access Token,
and use that for the checkout and git commands. Here's to hoping
there's an alternative in the near future.