Add Action to accept dispatches to update
I'm setting up the terraform-provider-njalla repository to send a
repository dispatch whenever a release is published.

This repository will take in that event, and use the tag from the
terraform-provider-njalla to update the PKGBUILD and push those
changes (which will then trigger the tag and release workflows from
this repository, ultimately tagging the Arch Linux package version
change, and releasing the compiled Arch Linux package for the new
version).